# backend/docs/server/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Server for CI/CD application developed using Go (Gin, Gorm).",
        "title": "CI/CD Application - Server API",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/workers": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workers"
                ],
                "summary": "Get all workers",
                "operationId": "all-workers",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Order by field",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "description": "Filter",
                        "name": "worker",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/model.WorkerCore"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of workers",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.WorkerShort"
                            }
                        }
                    },
                    "400": {
                        "description": "Error in request",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    },
                    "404": {
                        "description": "No records found",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workers"
                ],
                "summary": "Create new worker",
                "operationId": "create-worker",
                "parameters": [
                    {
                        "description": "New worker entry",
                        "name": "worker",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.WorkerCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Newly created worker",
                        "schema": {
                            "$ref": "#/definitions/model.Worker"
                        }
                    },
                    "400": {
                        "description": "Error in params",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    },
                    "501": {
                        "description": "Endpoint not implemented",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    }
                }
            }
        },
        "/workers/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workers"
                ],
                "summary": "Get the single worker",
                "operationId": "single-worker",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Worker ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Requested worker",
                        "schema": {
                            "$ref": "#/definitions/model.Worker"
                        }
                    },
                    "400": {
                        "description": "Error in params",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    },
                    "404": {
                        "description": "No record found",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workers"
                ],
                "summary": "Update worker",
                "operationId": "update-worker",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Worker ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated worker entry",
                        "name": "worker",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.WorkerCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated worker",
                        "schema": {
                            "$ref": "#/definitions/model.Worker"
                        }
                    },
                    "400": {
                        "description": "Error in params",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    },
                    "501": {
                        "description": "Endpoint not implemented",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workers"
                ],
                "summary": "Delete worker",
                "operationId": "delete-worker",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Worker ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete message",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    },
                    "400": {
                        "description": "Error in params",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    },
                    "501": {
                        "description": "Endpoint not implemented",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Build": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "number": {
                    "type": "integer"
                },
                "pipeline_id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/model.BuildStatus"
                },
                "steps": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.BuildStep"
                    }
                },
                "tags": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "worker_id": {
                    "type": "string"
                }
            }
        },
...
        "model.BuildLog": {
            "type": "object",
            "properties": {
                "build_step_id": {
                    "type": "string"
                },
                "command": {
                    "type": "string"
                },
                "output": {
                    "type": "string"
                }
            }
        },
...
        "model.BuildStatus": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3
            ],
            "x-enum-varnames": [
                "BuildRunning",
                "BuildSuccessful",
                "BuildFailed",
                "BuildCanceled"
            ]
        },
        "model.BuildStep": {
            "type": "object",
            "properties": {
                "build_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "duration": {
                    "$ref": "#/definitions/time.Duration"
                },
                "id": {
                    "type": "string"
                },
                "logs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.BuildLog"
                    }
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "model.Worker": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "builds": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Build"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/model.WorkerStatus"
                },
                "system": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/model.WorkerType"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "model.WorkerCore": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/model.WorkerStatus"
                },
                "system": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/model.WorkerType"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "model.WorkerCreate": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "private_key": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/model.WorkerStatus"
                },
                "system": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/model.WorkerType"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "model.WorkerShort": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/model.WorkerStatus"
                },
                "system": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/model.WorkerType"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "model.WorkerStatus": {
            "type": "integer",
            "enum": [
                0,
                1,
                2
            ],
            "x-enum-varnames": [
                "WorkerIdle",
                "WorkerUsed",
                "WorkerUnreachable"
            ]
        },
        "model.WorkerType": {
            "type": "integer",
            "enum": [
                0,
                1
            ],
            "x-enum-varnames": [
                "WorkerStatic",
                "WorkerDockerHost"
            ]
        },
        "time.Duration": {
            "type": "integer",
            "enum": [
                1,
                1000,
                1000000,
                1000000000
            ],
            "x-enum-varnames": [
                "Nanosecond",
                "Microsecond",
                "Millisecond",
                "Second"
            ]
        },
        "util.Message": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}